Enable tsconfig rule noImplicitReturns: true and deal with resulting errors
