+++ tsconfig.json
@@ -34,7 +34,7 @@
         "noUnusedLocals": true,
         /* Report errors on unused locals. */
         // "noUnusedParameters": true,            /* Report errors on unused parameters. */
-        // "noImplicitReturns": true,             /* Report error when not all code paths in function return a value. */
+        "noImplicitReturns": true,             /* Report error when not all code paths in function return a value. */
         // "noFallthroughCasesInSwitch": true,    /* Report errors for fallthrough cases in switch statement. */
 
         /* Module Resolution Options */
@@ -70,5 +70,6 @@
         // "experimentalDecorators": true,        /* Enables experimental support for ES7 decorators. */
         // "emitDecoratorMetadata": true,         /* Enables experimental support for emitting type metadata for decorators. */
     },
-    "exclude": ["cypress"]
+	"include": ["src/**/*"],
+    "exclude": ["cypress", "src/**/*.test.ts", "src/**/*.test.tsx"]
 }
